+++ LanguageExt.Megaparsec/PosState.cs
@@ -2,7 +2,7 @@ namespace LanguageExt.Megaparsec;
 
 public readonly record struct PosState<S>(
     S Input, 
-    int Offset, 
+    long Offset, 
     SourcePos SourcePos, 
     int TabWidth, 
     string LinePrefix);
