+++ LanguageExt.Megaparsec/State.cs
@@ -2,6 +2,6 @@ namespace LanguageExt.Megaparsec;
 
 public readonly record struct State<S, T, E>(
     S Input,
-    int Offset,
+    long Offset,
     PosState<S> PosState,
     Seq<ParseError<T, E>> ParseErrors);
